Reorganize the latex converter code
